Move menu button field to shared lib
* Menu button field was moved to shared library
