+++ apps/admin-gui/src/app/facilities/pages/resource-detail-page/resource-overview/resource-overview.component.ts
@@ -1,6 +1,6 @@
 import {Component, HostBinding, OnInit} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
-import {MenuItem} from '../../../../shared/models/MenuItem';
+import {MenuItem} from '../../../../../../../../libs/perun/models/src/lib/MenuItem';
 import { Resource, ResourcesManagerService } from '@perun-web-apps/perun/openapi';
 
 @Component({
