+++ apps/admin-gui/src/app/shared/components/user-detail-page/user-settings/user-settings-overview/user-settings-overview.component.ts
@@ -1,5 +1,5 @@
 import {Component, HostBinding, OnInit} from '@angular/core';
-import {MenuItem} from '../../../../models/MenuItem';
+import {MenuItem} from '../../../../../../../../../libs/perun/models/src/lib/MenuItem';
 
 @Component({
   selector: 'app-user-settings-overview',
