+++ apps/admin-gui/src/app/vos/pages/group-detail-page/group-overview/group-overview.component.ts
@@ -1,5 +1,5 @@
 import { Component, HostBinding, OnInit } from '@angular/core';
-import { MenuItem } from '../../../../shared/models/MenuItem';
+import { MenuItem } from '../../../../../../../../libs/perun/models/src/lib/MenuItem';
 import { ActivatedRoute } from '@angular/router';
 import { Group, GroupsManagerService } from '@perun-web-apps/perun/openapi';
