+++ apps/admin-gui/src/app/vos/pages/group-detail-page/group-settings/group-settings-overview/group-settings-overview.component.ts
@@ -1,7 +1,7 @@
 import { Component, HostBinding, OnInit } from '@angular/core';
 import { SideMenuService } from '../../../../../core/services/common/side-menu.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { MenuItem } from '../../../../../shared/models/MenuItem';
+import { MenuItem } from '../../../../../../../../../libs/perun/models/src/lib/MenuItem';
 import { Group, GroupsManagerService, Vo, VosManagerService } from '@perun-web-apps/perun/openapi';
 
 @Component({
